Moved classes for termination of iterative algorithms to more specific namespaces
