+++ src/MGroup.LinearAlgebra/Iterative/PreconditionedConjugateGradient/PcgAlgorithmBase.cs
@@ -1,8 +1,9 @@
 using System;
+
 using MGroup.LinearAlgebra.Commons;
 using MGroup.LinearAlgebra.Exceptions;
 using MGroup.LinearAlgebra.Iterative.Preconditioning;
-using MGroup.LinearAlgebra.Iterative.Termination;
+using MGroup.LinearAlgebra.Iterative.Termination.Iterations;
 using MGroup.LinearAlgebra.Matrices;
 using MGroup.LinearAlgebra.Vectors;
