+++ src/MGroup.LinearAlgebra/Iterative/Termination/Convergence/RelativeSolutionConvergenceCriterion.cs
@@ -1,8 +1,9 @@
-namespace MGroup.LinearAlgebra.Iterative.Termination
+namespace MGroup.LinearAlgebra.Iterative.Termination.Convergence
 {
 	using System;
 	using System.Collections.Generic;
 	using System.Text;
+
 	using MGroup.LinearAlgebra.Vectors;
 
 	/// <summary>
@@ -14,8 +15,8 @@ namespace MGroup.LinearAlgebra.Iterative.Termination
 		{
 			//TODO: The next can be optimized to not create a new vector (using SubtractIntoThis) in some cases.
 			//		E.g. in Gauss-Seidel the previousSolution vector is no longer necessary and can be overwritten.
-			double num = previousSolution.Subtract(currentSolution).Norm2();
-			double den = currentSolution.Norm2();
+			var num = previousSolution.Subtract(currentSolution).Norm2();
+			var den = currentSolution.Norm2();
 			if (den != 0)
 			{
 				return num / den;
