--- src/MGroup.LinearAlgebra/Iterative/Termination/IStagnationCriterion.cs
@@ -1,16 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-//TODO: Combine this with convergence criterion
-namespace MGroup.LinearAlgebra.Iterative.Termination
-{
-    public interface IStagnationCriterion
-    {
-        bool HasStagnated();
-
-        void StoreInitialError(double initialError);
-
-        void StoreNewError(double currentError);
-    }
-}
