--- src/MGroup.LinearAlgebra/Iterative/Termination/NullStagnationCriterion.cs
@@ -1,16 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace MGroup.LinearAlgebra.Iterative.Termination
-{
-    public class NullStagnationCriterion : IStagnationCriterion
-    {
-        public bool HasStagnated() => false;
-
-        public void StoreInitialError(double initialError) { }
-
-        public void StoreNewError(double currentError) { }
-
-    }
-}
