+++ src/MGroup.LinearAlgebra/Iterative/Termination/Stegnation/IStagnationCriterion.cs
@@ -1,0 +1,16 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+//TODO: Combine this with convergence criterion
+namespace MGroup.LinearAlgebra.Iterative.Termination.Stegnation
+{
+	public interface IStagnationCriterion
+	{
+		bool HasStagnated();
+
+		void StoreInitialError(double initialError);
+
+		void StoreNewError(double currentError);
+	}
+}
