+++ src/MGroup.LinearAlgebra/Iterative/Termination/Stegnation/NullStagnationCriterion.cs
@@ -1,0 +1,16 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MGroup.LinearAlgebra.Iterative.Termination.Stegnation
+{
+	public class NullStagnationCriterion : IStagnationCriterion
+	{
+		public bool HasStagnated() => false;
+
+		public void StoreInitialError(double initialError) { }
+
+		public void StoreNewError(double currentError) { }
+
+	}
+}
