+++ tests/MGroup.LinearAlgebra.Tests/Iterative/PodAmg/PodAmgTests.cs
@@ -9,7 +9,8 @@ namespace MGroup.LinearAlgebra.Tests.Iterative.PodAmg
 	using MGroup.LinearAlgebra.Iterative.AlgebraicMultiGrid.PodAmg;
 	using MGroup.LinearAlgebra.Iterative.AlgebraicMultiGrid.Smoothing;
 	using MGroup.LinearAlgebra.Iterative.GaussSeidel;
-	using MGroup.LinearAlgebra.Iterative.Termination;
+	using MGroup.LinearAlgebra.Iterative.Termination.Convergence;
+	using MGroup.LinearAlgebra.Iterative.Termination.Iterations;
 	using MGroup.LinearAlgebra.Matrices;
 	using MGroup.LinearAlgebra.Tests.Utilities;
 	using MGroup.LinearAlgebra.Vectors;
